Trying to fix the incompatibility

+++ packages/Utils/src/DTK_Types.h
@@ -30,6 +30,6 @@ typedef double Coordinate;
 typedef int LocalOrdinal;
 
 //! Global ordinal typedef.
-typedef int64_t GlobalOrdinal;
+typedef long long GlobalOrdinal;
 
 #endif // DTK_TYPES_H
